Implement fixes from nomic audit
- Fix assertRevert
- Fix failing test in ProxyFactory
- Remove invalid test from Transactions
- Add tests for Initializable proxy
- Update comment on ImplementationProvider
