+++ packages/lib/package.json
@@ -59,7 +59,6 @@
     "lodash.isstring": "^4.0.1",
     "lodash.keys": "^4.2.0",
     "lodash.map": "^4.6.0",
-    "lodash.omit": "^4.5.0",
     "lodash.pick": "^4.4.0",
     "lodash.random": "^3.2.0",
     "lodash.reverse": "^4.0.1",
@@ -82,6 +81,7 @@
     "ganache-cli": "^6.4.0",
     "ganache-core": "^2.5.1",
     "lodash.foreach": "^4.5.0",
+    "lodash.omit": "^4.5.0",
     "lodash.times": "^4.3.2",
     "lodash.zipwith": "^4.2.0",
     "mock-dependency": "file:test/mocks/mock-dependency",
