+++ packages/lib/test/contracts/upgradeability/UpgradeabilityProxy.test.js
@@ -6,5 +6,9 @@ import shouldBehaveLikeUpgradeabilityProxy from './UpgradeabilityProxy.behaviour
 const UpgradeabilityProxy = artifacts.require('UpgradeabilityProxy')
 
 contract('UpgradeabilityProxy', ([_, proxyAdminOwner]) => {
-  shouldBehaveLikeUpgradeabilityProxy(UpgradeabilityProxy, undefined, proxyAdminOwner);
+  const createProxy = async function (implementation, _admin, initData, opts) {
+    return UpgradeabilityProxy.new(implementation, initData, opts);
+  };
+
+  shouldBehaveLikeUpgradeabilityProxy(createProxy, undefined, proxyAdminOwner);
 })
